Basic Function Completion of New Version

+++ app/src/main/java/io/taucoin/android/wallet/db/util/MigrationHelper.java
@@ -12,6 +12,7 @@ import org.greenrobot.greendao.database.Database;
 import org.greenrobot.greendao.database.StandardDatabase;
 import org.greenrobot.greendao.internal.DaoConfig;
 
+import java.lang.ref.WeakReference;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -31,9 +32,26 @@ public class MigrationHelper {
     private static final String SQLITE_MASTER = "sqlite_master";
     private static final String SQLITE_TEMP_MASTER = "sqlite_temp_master";
 
+    private static WeakReference<ReCreateAllTableListener> weakListener;
+
+    public interface ReCreateAllTableListener{
+        void onCreateAllTables(Database db, boolean ifNotExists);
+        void onDropAllTables(Database db, boolean ifExists);
+    }
+
     public static void migrate(SQLiteDatabase db, Class<? extends AbstractDao<?, ?>>... daoClasses) {
         printLog("【The Old Database Version】" + db.getVersion());
         Database database = new StandardDatabase(db);
+        migrate(database, daoClasses);
+    }
+
+    public static void migrate(SQLiteDatabase db, ReCreateAllTableListener listener, Class<? extends AbstractDao<?, ?>>... daoClasses) {
+        weakListener = new WeakReference<>(listener);
+        migrate(db, daoClasses);
+    }
+
+    public static void migrate(Database database, ReCreateAllTableListener listener, Class<? extends AbstractDao<?, ?>>... daoClasses) {
+        weakListener = new WeakReference<>(listener);
         migrate(database, daoClasses);
     }
 
@@ -42,9 +60,20 @@ public class MigrationHelper {
         generateTempTables(database, daoClasses);
         printLog("【Generate temp table】complete");
 
-        dropAllTables(database, true, daoClasses);
-        createAllTables(database, false, daoClasses);
+        ReCreateAllTableListener listener = null;
+        if (weakListener != null) {
+            listener = weakListener.get();
+        }
 
+        if (listener != null) {
+            listener.onDropAllTables(database, true);
+            printLog("【Drop all table by listener】");
+            listener.onCreateAllTables(database, false);
+            printLog("【Create all table by listener】");
+        } else {
+            dropAllTables(database, true, daoClasses);
+            createAllTables(database, false, daoClasses);
+        }
         printLog("【Restore data】start");
         restoreData(database, daoClasses);
         printLog("【Restore data】complete");
@@ -120,12 +149,12 @@ public class MigrationHelper {
 
     private static void dropAllTables(Database db, boolean ifExists, @NonNull Class<? extends AbstractDao<?, ?>>... daoClasses) {
         reflectMethod(db, "dropTable", ifExists, daoClasses);
-        printLog("【Drop all table】");
+        printLog("【Drop all table by reflect】");
     }
 
     private static void createAllTables(Database db, boolean ifNotExists, @NonNull Class<? extends AbstractDao<?, ?>>... daoClasses) {
         reflectMethod(db, "createTable", ifNotExists, daoClasses);
-        printLog("【Create all table】");
+        printLog("【Create all table by reflect】");
     }
 
     /**
@@ -161,22 +190,39 @@ public class MigrationHelper {
 
             try {
                 // get all columns from tempTable, take careful to use the columns list
-                List<String> columns = getColumns(db, tempTableName);
-                ArrayList<String> properties = new ArrayList<>(columns.size());
-                for (int j = 0; j < daoConfig.properties.length; j++) {
-                    String columnName = daoConfig.properties[j].columnName;
-                    if (columns.contains(columnName)) {
-                        properties.add(columnName);
+                List<TableInfo> newTableInfos = TableInfo.getTableInfo(db, tableName);
+                List<TableInfo> tempTableInfos = TableInfo.getTableInfo(db, tempTableName);
+                ArrayList<String> selectColumns = new ArrayList<>(newTableInfos.size());
+                ArrayList<String> intoColumns = new ArrayList<>(newTableInfos.size());
+                for (TableInfo tableInfo : tempTableInfos) {
+                    if (newTableInfos.contains(tableInfo)) {
+                        String column = '`' + tableInfo.name + '`';
+                        intoColumns.add(column);
+                        selectColumns.add(column);
                     }
                 }
-                if (properties.size() > 0) {
-                    final String columnSQL = TextUtils.join(",", properties);
+                // NOT NULL columns list
+                for (TableInfo tableInfo : newTableInfos) {
+                    if (tableInfo.notnull && !tempTableInfos.contains(tableInfo)) {
+                        String column = '`' + tableInfo.name + '`';
+                        intoColumns.add(column);
 
+                        String value;
+                        if (tableInfo.dfltValue != null) {
+                            value = "'" + tableInfo.dfltValue + "' AS ";
+                        } else {
+                            value = "'' AS ";
+                        }
+                        selectColumns.add(value + column);
+                    }
+                }
+
+                if (intoColumns.size() != 0) {
                     StringBuilder insertTableStringBuilder = new StringBuilder();
-                    insertTableStringBuilder.append("INSERT INTO ").append(tableName).append(" (");
-                    insertTableStringBuilder.append(columnSQL);
+                    insertTableStringBuilder.append("REPLACE INTO ").append(tableName).append(" (");
+                    insertTableStringBuilder.append(TextUtils.join(",", intoColumns));
                     insertTableStringBuilder.append(") SELECT ");
-                    insertTableStringBuilder.append(columnSQL);
+                    insertTableStringBuilder.append(TextUtils.join(",", selectColumns));
                     insertTableStringBuilder.append(" FROM ").append(tempTableName).append(";");
                     db.execSQL(insertTableStringBuilder.toString());
                     printLog("【Restore data】 to " + tableName);
@@ -216,5 +262,57 @@ public class MigrationHelper {
         }
     }
 
+    private static class TableInfo {
+        int cid;
+        String name;
+        String type;
+        boolean notnull;
+        String dfltValue;
+        boolean pk;
+
+        @Override
+        public boolean equals(Object o) {
+            return this == o
+                    || o != null
+                    && getClass() == o.getClass()
+                    && name.equals(((TableInfo) o).name);
+        }
+
+        @Override
+        public String toString() {
+            return "TableInfo{" +
+                    "cid=" + cid +
+                    ", name='" + name + '\'' +
+                    ", type='" + type + '\'' +
+                    ", notnull=" + notnull +
+                    ", dfltValue='" + dfltValue + '\'' +
+                    ", pk=" + pk +
+                    '}';
+        }
+
+        private static List<TableInfo> getTableInfo(Database db, String tableName) {
+            String sql = "PRAGMA table_info(" + tableName + ")";
+            printLog(sql);
+            Cursor cursor = db.rawQuery(sql, null);
+            if (cursor == null)
+                return new ArrayList<>();
+
+            TableInfo tableInfo;
+            List<TableInfo> tableInfos = new ArrayList<>();
+            while (cursor.moveToNext()) {
+                tableInfo = new TableInfo();
+                tableInfo.cid = cursor.getInt(0);
+                tableInfo.name = cursor.getString(1);
+                tableInfo.type = cursor.getString(2);
+                tableInfo.notnull = cursor.getInt(3) == 1;
+                tableInfo.dfltValue = cursor.getString(4);
+                tableInfo.pk = cursor.getInt(5) == 1;
+                tableInfos.add(tableInfo);
+                // printLog(tableName + "：" + tableInfo);
+            }
+            cursor.close();
+            return tableInfos;
+        }
+    }
 }
